Restructure src folder
The routes folder contains the Express routers for the application. Contains the API endpoints and CRUD methods.

The services folder contains the main logic for the application.

The models folder is used to define the structure of documents in a MongoDB collection, it contains the Mongoose models for the application.

The interfaces folder is used to define the structure of objects in the application's code.